refactor(api): split rest from user service and delete it
