+++ packages/api/src/modules/user/@types/UserServiceError.ts
@@ -1,0 +1,4 @@
+export interface UserServiceError {
+    message: string;
+    code: number;
+}
