+++ packages/api/src/modules/user/services/consumer/user.consumer.service.test.ts
@@ -1,7 +1,6 @@
 import { USER_EMAIL } from "../../../../../tests/__helpers__/userHelper";
 import { InternalServerError } from "../../../../shared/errors/httpErrors";
 import { CONSUMER_JWT_PAYLOAD, CONSUMER_USER, USER_WITHOUT_SECRET } from "../../__fixtures__/user.fixture";
-import { UserServiceErrors } from "../../user.service";
 import userConsumerService from "./user.consumer.service";
 import userCrudService from "../crud/user.crud.service";
 jest.mock("../crud/user.crud.service");
@@ -10,6 +9,7 @@ import userAuthService from "../auth/user.auth.service";
 jest.mock("../auth/user.auth.service");
 const mockedUserAuthService = jest.mocked(userAuthService);
 import consumerTokenRepository from "../../repositories/consumer-token.repository";
+import { UserServiceErrors } from "../../user.enum";
 jest.mock("../../repositories/consumer-token.repository");
 const mockedConsumerTokenRepository = jest.mocked(consumerTokenRepository);
