+++ packages/api/src/modules/user/services/crud/user.crud.service.test.ts
@@ -1,7 +1,7 @@
 import userCrudService from "./user.crud.service";
 import userRepository from "../../repositories/user.repository";
 import { USER_EMAIL } from "../../../../../tests/__helpers__/userHelper";
-import { SIGNED_TOKEN, USER_WITHOUT_SECRET } from "../../__fixtures__/user.fixture";
+import { SIGNED_TOKEN, USER_DBO, USER_WITHOUT_SECRET } from "../../__fixtures__/user.fixture";
 import bcrypt from "bcrypt";
 jest.mock("bcrypt");
 jest.mock("../../repositories/user.repository");
@@ -33,6 +33,8 @@ jest.mock("../../../notify/notify.service", () => ({
     notify: jest.fn(),
 }));
 const mockedNotifyService = jest.mocked(notifyService);
+import * as repositoryHelper from "../../../../shared/helpers/RepositoryHelper";
+jest.mock("../../../../shared/helpers/RepositoryHelper");
 
 describe("user crud service", () => {
     describe("find", () => {
@@ -241,6 +243,28 @@ describe("user crud service", () => {
             mockCreateUser.mockImplementationOnce(async () => expected as UserDto);
             const actual = await userCrudService.signup({ email: USER_EMAIL });
             expect(actual).toEqual(expected);
+        });
+    });
+
+    describe("getUserWithoutSecret", () => {
+        const EMAIL = "user@mail.fr";
+
+        it("gets user from repo", async () => {
+            jest.mocked(userRepository.getUserWithSecretsByEmail).mockResolvedValueOnce(USER_DBO);
+            await userCrudService.getUserWithoutSecret(EMAIL);
+            expect(userRepository.getUserWithSecretsByEmail).toHaveBeenCalledWith(EMAIL);
+        });
+
+        it("should call removeSecrets()", async () => {
+            jest.mocked(userRepository.getUserWithSecretsByEmail).mockResolvedValueOnce(USER_DBO);
+            const actual = await userCrudService.getUserWithoutSecret(EMAIL);
+            expect(repositoryHelper.removeSecrets).toHaveBeenCalledTimes(1);
+        });
+
+        it("throws not found if noe found", async () => {
+            jest.mocked(userRepository.getUserWithSecretsByEmail).mockResolvedValueOnce(null);
+            const test = () => userCrudService.getUserWithoutSecret(EMAIL);
+            await expect(test).rejects.toMatchInlineSnapshot(`[Error: User not found]`);
         });
     });
 });
