+++ packages/api/src/modules/user/services/roles/user.roles.service.ts
@@ -2,7 +2,7 @@ import { UserDto } from "dto";
 import { RoleEnum } from "../../../../@enums/Roles";
 import userRepository from "../../repositories/user.repository";
 import { BadRequestError, InternalServerError } from "../../../../shared/errors/httpErrors";
-import { UserServiceErrors } from "../../user.service";
+import { UserServiceErrors } from "../../user.enum";
 
 export class UserRolesService {
     public getRoles(user: UserDto) {
