+++ packages/api/src/modules/user/user.enum.ts
@@ -1,0 +1,17 @@
+export enum UserServiceErrors {
+    LOGIN_WRONG_PASSWORD_MATCH,
+    LOGIN_UPDATE_JWT_FAIL,
+    USER_NOT_ACTIVE,
+    CREATE_INVALID_EMAIL,
+    CREATE_USER_ALREADY_EXISTS,
+    CREATE_USER_WRONG,
+    FORMAT_PASSWORD_INVALID,
+    USER_NOT_FOUND,
+    ROLE_NOT_FOUND,
+    RESET_TOKEN_NOT_FOUND,
+    RESET_TOKEN_EXPIRED,
+    CREATE_RESET_PASSWORD_WRONG,
+    CREATE_EMAIL_GOUV,
+    CREATE_CONSUMER_TOKEN,
+    USER_TOKEN_EXPIRED,
+}
