--- packages/api/src/modules/user/user.service.ts
@@ -1,91 +1,0 @@
-import { ObjectId } from "mongodb";
-
-import dedent from "dedent";
-import { SignupErrorCodes, UserWithResetTokenDto, UserDto, FutureUserDto } from "dto";
-import { RoleEnum } from "../../@enums/Roles";
-import { DefaultObject } from "../../@types";
-import { JWT_EXPIRES_TIME } from "../../configurations/jwt.conf";
-import { BadRequestError, InternalServerError, NotFoundError } from "../../shared/errors/httpErrors";
-import { NotificationType } from "../notify/@types/NotificationType";
-import notifyService from "../notify/notify.service";
-import { removeSecrets } from "../../shared/helpers/RepositoryHelper";
-import { FRONT_OFFICE_URL } from "../../configurations/front.conf";
-import { ConsumerToken } from "./entities/ConsumerToken";
-import consumerTokenRepository from "./repositories/consumer-token.repository";
-import userResetRepository from "./repositories/user-reset.repository";
-import { UserNotPersisted } from "./repositories/dbo/UserDbo";
-import userAuthService from "./services/auth/user.auth.service";
-import userRepository from "./repositories/user.repository";
-import { DEFAULT_PWD } from "./user.constant";
-import userCheckService from "./services/check/user.check.service";
-import userStatsService from "./services/stats/user.stats.service";
-import userActivationService from "./services/activation/user.activation.service";
-import userCrudService from "./services/crud/user.crud.service";
-
-export enum UserServiceErrors {
-    LOGIN_WRONG_PASSWORD_MATCH,
-    LOGIN_UPDATE_JWT_FAIL,
-    USER_NOT_ACTIVE,
-    CREATE_INVALID_EMAIL,
-    CREATE_USER_ALREADY_EXISTS,
-    CREATE_USER_WRONG,
-    FORMAT_PASSWORD_INVALID,
-    USER_NOT_FOUND,
-    ROLE_NOT_FOUND,
-    RESET_TOKEN_NOT_FOUND,
-    RESET_TOKEN_EXPIRED,
-    CREATE_RESET_PASSWORD_WRONG,
-    CREATE_EMAIL_GOUV,
-    CREATE_CONSUMER_TOKEN,
-    USER_TOKEN_EXPIRED,
-}
-
-export interface UserServiceError {
-    message: string;
-    code: number;
-}
-
-export class UserService {
-    public static RESET_TIMEOUT = 1000 * 60 * 60 * 24 * 10; // 10 days in ms
-    public static PASSWORD_VALIDATOR_MESSAGE = dedent`Password is too weak, please use this rules:
-        At least one digit [0-9]
-        At least one lowercase character [a-z]
-        At least one uppercase character [A-Z]
-        At least one special character [*.!@#$%^&(){}[]:;<>,.?/~_+-=|\\]
-        At least 8 characters in length, but no more than 32.`;
-
-    public static CONSUMER_TOKEN_PROP = "isConsumerToken";
-
-    async findConsumerToken(userId: ObjectId): Promise<string> {
-        const token = await consumerTokenRepository.findToken(userId);
-        if (!token) {
-            throw new NotFoundError("Aucun token d'authentification n'a été trouvé");
-        }
-        return token;
-    }
-
-    async findAnonymizedUsers(query: DefaultObject = {}) {
-        const users = await userCrudService.find(query);
-
-        return users.map(user => {
-            return {
-                ...user,
-                _id: user._id.toString(),
-                email: undefined,
-                firstName: undefined,
-                lastName: undefined,
-                phoneNumber: undefined,
-            };
-        });
-    }
-
-    async getUserWithoutSecret(email: string) {
-        const withSecrets = await userRepository.getUserWithSecretsByEmail(email);
-        if (!withSecrets) throw new NotFoundError("User not found");
-        return removeSecrets(withSecrets);
-    }
-}
-
-const userService = new UserService();
-
-export default userService;
