+++ packages/api/tests/__helpers__/userHelper.ts
@@ -1,7 +1,6 @@
 import { RoleEnum } from "../../src/@enums/Roles";
 import userActivationService from "../../src/modules/user/services/activation/user.activation.service";
 import userCrudService from "../../src/modules/user/services/crud/user.crud.service";
-import userService from "../../src/modules/user/user.service";
 
 export const ADMIN_EMAIL = "admin@beta.gouv.fr";
 export const USER_EMAIL = "user@beta.gouv.fr";
