+++ packages/api/tests/modules/user/authentification.spec.ts
@@ -1,11 +1,12 @@
 import request = require("supertest");
-import userService, { UserService } from "../../../src/modules/user/user.service";
 import { AgentTypeEnum, ResetPasswordErrorCodes } from "dto";
 import { createAndActiveUser, createUser, DEFAULT_PASSWORD, USER_EMAIL } from "../../__helpers__/userHelper";
 import { createResetToken } from "../../__helpers__/resetTokenHelper";
 import userResetRepository from "../../../src/modules/user/repositories/user-reset.repository";
 import notifyService from "../../../src/modules/notify/notify.service";
-import userActivationService from "../../../src/modules/user/services/activation/user.activation.service";
+import userActivationService, {
+    UserActivationService,
+} from "../../../src/modules/user/services/activation/user.activation.service";
 import userCrudService from "../../../src/modules/user/services/crud/user.crud.service";
 
 const g = global as unknown as { app: unknown };
@@ -104,8 +105,8 @@ describe("AuthentificationController, /auth", () => {
 
             const userReset = await userResetRepository.findOneByUserId(user._id);
 
-            const oldResetTimout = UserService.RESET_TIMEOUT;
-            UserService.RESET_TIMEOUT = 0;
+            const oldResetTimout = UserActivationService.RESET_TIMEOUT;
+            UserActivationService.RESET_TIMEOUT = 0;
 
             const response = await request(g.app)
                 .post("/auth/reset-password")
@@ -114,14 +115,14 @@ describe("AuthentificationController, /auth", () => {
                     token: userReset?.token,
                 })
                 .set("Accept", "application/json");
-            UserService.RESET_TIMEOUT = oldResetTimout;
+            UserActivationService.RESET_TIMEOUT = oldResetTimout;
 
             expect(response.statusCode).toBe(400);
             expect(response.body).toMatchObject({
                 code: ResetPasswordErrorCodes.RESET_TOKEN_EXPIRED,
             });
 
-            UserService.RESET_TIMEOUT = oldResetTimout;
+            UserActivationService.RESET_TIMEOUT = oldResetTimout;
         });
     });
